bot/requirements.txt
python-telegram-bot
kafka-python
